create constant_src.rs
